+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 
 ### Added
 
+* `Add Faker.Beer.brand/0` [[@vbrazo][]]
 * `Add Faker.Food.sushi/0` [[@vbrazo][]]
 * `Faker.Team.PtBr` [[@vbrazo][]]
 * Tests covered by actual values not just checking that result is a binary [[@igas][]]